disfunctional attempt to incorporate autoGrab into robot. DO NOT MERGE FROM MY BRANCH. robot is broken

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/powerplay/CDS.java
@@ -113,7 +113,7 @@ public class CDS implements SubsystemManager {
         int red = getRed();
         int green = getGreen();
         int blue = getGreen();
-        int dist = color.getDistance(DistanceUnit.CM);
+        double dist = color.getDistance(DistanceUnit.CM);
         return ("red - "+ red+"\ngreen - "+green+"\nblue - "+blue+"\ndistance (cm)"+dist);
     }
     /**
